recoil update && stat

+++ Assets/_Scripts/ShakeCamera.cs
@@ -71,7 +71,7 @@ public class ShakeCamera : MonoBehaviour
 
         if (Input.GetKey(KeyCode.Mouse1))
         {
-            int handling = myGunHandler.actualGun.stats.GetStat(StatNames.Handling);
+            float handling = myGunHandler.actualGun.stats.GetStat(StatNames.Handling);
             actualAimPos = Vector3.Lerp(actualAimPos,aimPos - GunHandler.sightPosition,(soft + handling )* Time.deltaTime);
 
             hands.transform.localPosition = actualAimPos + temp1;
